+++ MuckSaveGame/Network/ServerMethods.cs
@@ -5,6 +5,7 @@
 	using Steamworks;
 	using System.Collections.Generic;
 	using System.Reflection;
+	using UnityEngine;
 
 	[HarmonyPatch]
 	internal class ServerMethods